feat(user-info-edit): add resetBasicState util to reset zustand and server state except nickname-#179
